xu ly token o buoc dang nhap

+++ src/main/java/org/example/lv_backend/entity/Chapter.java
@@ -21,7 +21,7 @@ public class Chapter {
         @GeneratedValue(strategy = GenerationType.IDENTITY)
         private Long id;
 
-        private Integer chapterNumber;
+        private Long chapterNumber;
 
         private String title;
