+++ src/main/java/org/example/lv_backend/repository/UserRepository.java
@@ -1,5 +1,6 @@
 package org.example.lv_backend.repository;
 
+import org.example.lv_backend.dto.response.user.UserResponse;
 import org.example.lv_backend.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -8,12 +9,13 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User,Integer> {
+public interface UserRepository extends JpaRepository<User,Long> {
     Optional<User> findByEmail(String email);
 
     boolean existsByName(String username);
 
     boolean existsByEmail(String email);
+
 
 }
